Make test.dart strong mode clean!
R=whesse@google.com

Review-Url: https://codereview.chromium.org/2855883008 .

+++ tools/testing/dart/test_configurations.dart
@@ -62,7 +62,7 @@ Future testConfigurations(List<Map> configurations) async {
   var startTime = new DateTime.now();
   // Extract global options from first configuration.
   var firstConf = configurations[0];
-  var maxProcesses = firstConf['tasks'];
+  var maxProcesses = firstConf['tasks'] as int;
   var progressIndicator = firstConf['progress'];
   BuildbotProgressIndicator.stepName = firstConf['step_name'];
   var verbose = firstConf['verbose'];
